+++ .vscode/settings.json
@@ -1,11 +1,9 @@
 {
-  "editor.formatOnSave":  false,
+  "editor.formatOnSave": true,
   "eslint.autoFixOnSave": true,
   "eslint.validate": [
     "javascript",
-    "javascriptreact",
-    { "language": "typescript", "autoFix": true },
-    { "language": "typescriptreact", "autoFix": true }
+    { "language": "typescript", "autoFix": true }
   ],
   "files.exclude": {
     "yarn.lock": true,
