+++ src/index.ts
@@ -1,7 +1,6 @@
 #!/usr/bin/env node
 
-import { initSteps } from './steps';
-import { printHandledError } from './HandledError';
+import { initSteps } from './steps/steps';
 import { getOptions } from './options/options';
 
 async function init() {
@@ -9,7 +8,12 @@ async function init() {
     const options = await getOptions(process.argv);
     return initSteps(options);
   } catch (e) {
-    printHandledError(e, { rethrow: false });
+    if (e.name === 'HandledError') {
+      console.error(e.message);
+    } else {
+      console.error(e);
+    }
+
     process.exit(1);
   }
 }
