+++ src/services/HandledError.ts
@@ -1,0 +1,7 @@
+export class HandledError extends Error {
+  constructor(message: string) {
+    super(message);
+    Error.captureStackTrace(this, HandledError);
+    this.name = 'HandledError';
+  }
+}
